Add files via upload
컬러 미선택시 널포인트 익셉션

+++ src/FigureDialog.java
@@ -48,7 +48,7 @@ public class FigureDialog extends JDialog {
 		JRadioButton chooserButton;
 		
 		Color color;
-		int thickness;
+		float thickness;
 		JDialog dialog;
 		DrawerView view;
