+++ src/OnePointFigure.java
@@ -7,11 +7,11 @@ abstract public class OnePointFigure extends Figure{
 	protected int x1;
 	protected int y1;
 	
-	OnePointFigure(Color color, int thickness) {
+	OnePointFigure(Color color, float thickness) {
 		super(color, thickness);
 	}
 
-	OnePointFigure(Color color, int thickness, int x, int y) {
+	OnePointFigure(Color color, float thickness, int x, int y) {
 		super(color, thickness);
 		x1 = x;
 		y1 = y;
